+++ FbonizziMonoGame/FbonizziMonoGame/AfterGameManagement/ProtipsShower.cs
@@ -1,4 +1,4 @@
-﻿using FbonizziMonoGame.Drawing;
+﻿using FbonizziMonoGame.Sprites;
 using FbonizziMonoGame.Extensions;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
