+++ FbonizziMonoGame/FbonizziMonoGame/Extensions/SpriteBatchExtensions.cs
@@ -1,4 +1,4 @@
-﻿using FbonizziMonoGame.Drawing;
+﻿using FbonizziMonoGame.Sprites;
 using FbonizziMonoGame.Sprites;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
