+++ FbonizziMonoGame/FbonizziMonoGame/Sprites/AnimationsManager.cs
@@ -1,26 +1,49 @@
 ﻿using FbonizziMonoGame.Extensions;
+using FbonizziMonoGame.Sprites;
 using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Collections.Generic;
 
-namespace FbonizziMonoGame.Drawing
+namespace FbonizziMonoGame.Sprites
 {
+    /// <summary>
+    /// It contains a collection of named animations and permits to play, stop, resume them
+    /// </summary>
     public class AnimationsManager
     { 
         private IDictionary<string, SpriteAnimation> _animations;
+
+        /// <summary>
+        /// Returns the current animation name
+        /// </summary>
         public string CurrentAnimationName { get; private set; }
+
+        /// <summary>
+        /// It returns the current sprite animation
+        /// </summary>
         public SpriteAnimation CurrentAnimation
             => _animations[CurrentAnimationName];
 
+        /// <summary>
+        /// Constructs an empty animation manager
+        /// </summary>
         public AnimationsManager()
         {
             _animations = new Dictionary<string, SpriteAnimation>();
         }
 
+        /// <summary>
+        /// Adds an animation to the manager
+        /// </summary>
+        /// <param name="animationKey"></param>
+        /// <param name="animation"></param>
+        /// <returns></returns>
         public AnimationsManager AddAnimation(
            string animationKey,
            SpriteAnimation animation)
         {
+            if (animationKey == null)
+                throw new ArgumentNullException(nameof(animationKey));
             if (animation == null)
                 throw new ArgumentNullException(nameof(animation));
 
@@ -28,6 +51,10 @@ namespace FbonizziMonoGame.Drawing
             return this;
         }
 
+        /// <summary>
+        /// Play an animation given its name
+        /// </summary>
+        /// <param name="animationKey"></param>
         public void PlayAnimation(string animationKey)
         {
             if (CurrentAnimationName == animationKey)
@@ -37,12 +64,26 @@ namespace FbonizziMonoGame.Drawing
             _animations[CurrentAnimationName].Play();
         }
 
+        /// <summary>
+        /// Returns a <see cref="SpriteAnimation"/> given its name
+        /// </summary>
+        /// <param name="animationName"></param>
+        /// <returns></returns>
         public SpriteAnimation GetAnimation(string animationName)
             => _animations[animationName];
 
+        /// <summary>
+        /// Manages the current animation logic
+        /// </summary>
+        /// <param name="elapsed"></param>
         public void Update(TimeSpan elapsed)
             => _animations[CurrentAnimationName].Update(elapsed);
 
+        /// <summary>
+        /// Draws the current animation frame
+        /// </summary>
+        /// <param name="spriteBatch"></param>
+        /// <param name="spatialInfos"></param>
         public void Draw(
             SpriteBatch spriteBatch,
             DrawingInfos spatialInfos)
